+++ src/CurrentProcess.php
@@ -6,8 +6,10 @@ namespace Innmind\OperatingSystem;
 use Innmind\OperatingSystem\CurrentProcess\Signals;
 use Innmind\Server\Control\Server\Process\Pid;
 use Innmind\Server\Status\Server\Memory\Bytes;
-use Innmind\TimeContinuum\Period;
-use Innmind\TimeWarp\Halt;
+use Innmind\Time\{
+    Period,
+    Halt,
+};
 use Innmind\Signals\Handler;
 use Innmind\Immutable\{
     Attempt,