+++ tests/CurrentProcessTest.php
@@ -11,8 +11,10 @@ use Innmind\OperatingSystem\{
 };
 use Innmind\Server\Control\Server\Process\Pid;
 use Innmind\Server\Status\Server\Memory\Bytes;
-use Innmind\TimeContinuum\Period;
-use Innmind\TimeWarp\Halt;
+use Innmind\Time\{
+    Period,
+    Halt,
+};
 use Innmind\Signals\{
     Handler,
     Signal,