+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - `Innmind\OperatingSystem\Filesystem::mount()` now longer automatically create the directory
 - `Innmind\OperatingSystem\CurrentProcess\Signals::listen()` now returns an `Innmind\Immutable\Attempt`
 - `Innmind\OperatingSystem\CurrentProcess\Signals::remove()` now returns an `Innmind\Immutable\Attempt`
+- Requires `innmind/time:~1.0`
 
 ## 6.2.0 - 2025-08-10